fix: type-o after renaming
