+++ src/components/index.js
@@ -45,7 +45,7 @@ import DpTextArea from './DpTextArea/DpTextArea'
 import DpTimePicker from './DpTimePicker/DpTimePicker'
 import DpToggle from './DpToggle/DpToggle'
 import DpToggleForm from './DpToggleForm/DpToggleForm'
-import DpTooltip from './DpTootip/DpTooltip.vue'
+import DpTooltip from './DpTooltip/DpTooltip'
 import DpTooltipIcon from './DpTooltipIcon/DpTooltipIcon'
 import DpTreeList from './DpTreeList/DpTreeList'
